add forceGlobal to Application, and NettyServer.createServer variant using it
Also support a NettyServer.createServer that contains the user.dir logic from main().

The overall goal is to allow creating a NettyServer that doesn't use a
global singleton or take over your main() and exit, for use in tests
for example.

The code here could be cleaner if back compat was not a concern; the patch
is trying to keep the previous constructor signatures in the ABI, so
can't just add forceGlobal:Option[GlobalSettings]=None.